Rename module from datacollection to datacapture
