+++ datacapture/proguard-rules.pro
