+++ datacapture/src/main/AndroidManifest.xml
