+++ datacapture/src/main/java/com/google/android/fhir/datacapture/QuestionnaireFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacollection
+package com.google.android.fhir.datacapture
 
 import android.os.Bundle
 import android.text.Editable
